requirements.txt
Django==5.1.1
django-unfold
django-eav2
django-phonenumber-field[phonenumbers]